update unit test and add mapper to dao and model api

+++ service/src/main/java/training/peopleandcars/services/CarServiceImpl.java
@@ -1,11 +1,14 @@
 package training.peopleandcars.services;
 
 
-import org.modelmapper.ModelMapper;
-import training.peopleandcars.modelapi.Car;
+import training.peopleandcars.exception.ModelNotFoundException;
+import training.peopleandcars.model.mapper.ConverterMapper;
+import training.peopleandcars.model.modelDao.CarDao;
+import training.peopleandcars.model.modelapi.Car;
 import training.peopleandcars.repository.CarRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,25 +16,42 @@ import java.util.Optional;
 @Service
 public class CarServiceImpl implements CarService {
     private CarRepository carRepository;
+    private RegistryService registryService;
+    private ConverterMapper converterMapper;
 
-    public CarServiceImpl(CarRepository carRepository) {
+    public CarServiceImpl(CarRepository carRepository,RegistryService registryService, ConverterMapper converterMapper) {
         this.carRepository = carRepository;
+        this.registryService =registryService;
+        this.converterMapper=converterMapper;
     }
 
     public Car save(Car car) {
-        return carRepository.save(car);
+        CarDao carDao= carRepository.save(converterMapper.toCarDao(car));
+        return converterMapper.toCar(carDao);
     }
 
     public List<Car> getAll() {
-        return carRepository.findAll();
+        List<Car> lstCar = new ArrayList<>();
+        List<CarDao> lstCarDao=carRepository.findAll();
+        for (CarDao carDao: lstCarDao
+             ) {
+            lstCar.add(converterMapper.toCar(carDao));
+        }
+
+        return lstCar;
     }
 
     public void delete(String vin) {
+        Car car = getById(vin);
+        if (registryService.getPeopleByCar(vin).size() > 0)
+            throw new ModelNotFoundException("The car is assigned, it cannot be deleted");
         carRepository.deleteById(vin);
     }
 
-    public Optional<Car> getById(String vin) {
-        return carRepository.findById(vin);
+    public Car getById(String vin) {
+        Optional<CarDao> carDao = carRepository.findById(vin);
+        if (carDao == null)
+            throw new ModelNotFoundException("Car not found");
+        return converterMapper.toCar(carDao.get());
     }
-
 }
